# api/services.swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Beer Catalog API",
    "version": "1.0.0",
    "description": "API for managing categories, beers and reviews"
  },
  "servers": [
    {
      "url": "http://localhost:80/api"
    }
  ],
  "tags": [
    {
      "name": "Categories",
      "description": "Все операции, связанные с категориями товаров"
    },
    {
      "name": "Beers",
      "description": "Операции, связанные с пивной продукцией"
    }
  ],
  "paths": {
    "/categories": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get all categories",
        "operationId": "getCategories",
        "responses": {
          "200": {
            "description": "List of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Category" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Categories"],
        "summary": "Create category",
        "operationId": "createCategory",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoryCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          }
        }
      }
    },

    "/categories/{id}": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get category by ID",
        "operationId": "getCategoryById",
        "parameters": [
          { "$ref": "#/components/parameters/CategoryId" }
        ],
        "responses": {
          "200": {
            "description": "Category",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          },
          "404": { "description": "Not found" }
        }
      },
      "patch": {
        "tags": ["Categories"],
        "summary": "Update category",
        "operationId": "updateCategory",
        "parameters": [
          { "$ref": "#/components/parameters/Id" }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CategoryUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Categories"],
        "summary": "Delete category",
        "operationId": "deleteCategory",
        "parameters": [
          { "$ref": "#/components/parameters/Id" }
        ],
        "responses": {
          "204": { "description": "Deleted" }
        }
      }
    },

    "/categories/parent/{id}": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get parent category",
        "operationId": "getParent",
        "parameters": [
          { "$ref": "#/components/parameters/Id" }
        ],
        "responses": {
          "200": {
            "description": "Parent",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Category" }
              }
            }
          }
        }
      }
    },

    "/categories/children/{id}": {
      "get": {
        "tags": ["Categories"],
        "summary": "Get child categories",
        "operationId": "getChildren",
        "parameters": [
          { "$ref": "#/components/parameters/Id" }
        ],
        "responses": {
          "200": {
            "description": "Children list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Category" }
                }
              }
            }
          }
        }
      }
    },

    "/beers": {
      "get": {
        "tags": ["Beers"],
        "summary": "Get all beers",
        "operationId": "getAllBeers",
        "parameters": [
          { "$ref": "#/components/parameters/Offset" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "Beers list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Beer" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Beers"],
        "summary": "Create beer",
        "operationId": "createBeer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BeerCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/beers/{id}": {
      "patch": {
        "tags": ["Beers"],
        "summary": "Update beer",
        "operationId": "updateBeer",
        "parameters": [
          { "$ref": "#/components/parameters/Id" }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BeerUpdate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated beer",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Beer" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Beers"],
        "summary": "Delete beer",
        "operationId": "deleteBeer",
        "parameters": [
          { "$ref": "#/components/parameters/Id" }
        ],
        "responses": {
          "200": {
            "description": "Delete status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/beers/reviews/{beer_id}": {
      "post": {
        "tags": ["Beers"],
        "summary": "Add review",
        "operationId": "addReview",
        "parameters": [
          { "$ref": "#/components/parameters/BeerId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ReviewCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" }
                  }
                }
              }
            }
          }
        }
      }
    },

    "/categories/beers/{category_id}": {
      "get": {
        "tags": ["Beers"],
        "summary": "Get beers by category",
        "operationId": "getBeersByCategory",
        "parameters": [
          { "$ref": "#/components/parameters/CategoryId" },
          { "$ref": "#/components/parameters/Offset" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "Beers list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Beer" }
                }
              }
            }
          }
        }
      }
    }
  },

  "components": {
    "parameters": {
      "CategoryId": {
        "name": "category_id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      },
      "Id": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      },
      "BeerId": {
        "name": "beer_id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer" }
      },

      "Offset": {
        "name": "offset",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 0,
          "minimum": 0
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 10,
          "minimum": 1,
          "maximum": 100
        }
      }
    },

    "schemas": {
      "Category": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "parent_id": { "type": "integer", "nullable": true }
        }
      },

      "CategoryCreate": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string" },
          "parent_id": { "type": "integer" }
        }
      },

      "CategoryUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "parent_id": { "type": "integer" }
        }
      },

      "Beer": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "abv": { "type": "number" },
          "ibu": { "type": "number" },
          "city": { "type": "string" },
          "country": { "type": "string" },
          "category": { "$ref": "#/components/schemas/Category" },
          "amount": {"type": "integer"},
          "units": {"type":  "string"},
          "type": { "type": "string" }
        }
      },

      "BeerCreate": {
        "type": "object",
        "required": ["name", "abv", "ibu", "city", "category", "type"],
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "abv": { "type": "number" },
          "ibu": { "type": "number" },
          "city": { "type": "string" },
          "country": { "type": "string" },
          "category": { "$ref": "#/components/schemas/Category" }, 
          "amount": {"type": "integer"},
          "units": {"type":  "string"},
          "type": { "type": "string" }
        }
      },

      "BeerUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "abv": { "type": "number" },
          "ibu": { "type": "number" },
          "city": { "type": "string" },
          "country": { "type": "string" },
          "category": { "$ref": "#/components/schemas/Category" },
          "amount": {"type": "integer"},
          "units": {"type":  "string"},
          "type": { "type": "string" }
        }
      },

      "BeerShort": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "rating": { "type": "number" },
          "description": { "type": "string" }
        }
      },

      "PaginatedBeers": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BeerShort"
            }
          },
          "offset": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "total_pages": { "type": "integer" },
          "has_next": { "type": "boolean" },
          "has_prev": { "type": "boolean" }
        }
      },

      "ReviewCreate": {
        "type": "object",
        "required": ["body", "rating"],
        "properties": {
          "body": { "type": "string" },
          "rating": { "type": "number" }
        }
      }
    }
  }
}